move ProgramPast to ffront.stages
