+++ src/gt4py/next/otf/stages.py
@@ -33,13 +33,6 @@ SettingT_co = TypeVar("SettingT_co", bound=languages.LanguageSettings, covariant
 
 
 @dataclasses.dataclass(frozen=True)
-class ProgramPast:
-    past_node: past.Program
-    closure_vars: dict[str, Any]
-    grid_type: Optional[common.GridType] = None
-
-
-@dataclasses.dataclass(frozen=True)
 class PastClosure:
     closure_vars: dict[str, Any]
     past_node: past.Program
